♾️ refactored infrastructure/ -> components/
Now all the resources are created by components from the
infrastructure/index.ts file. The resources are the same
as before.
