+++ deploy/infrastructure/index.ts
@@ -1,11 +1,26 @@
-import * as aws from "@pulumi/aws";
-import * as vpc from "./vpc";
-import * as alb from "./load_balancer";
-
+import * as pulumi from "@pulumi/pulumi";
 import { Cluster, Service } from "../components/cluster";
 import { Image } from "../components/image"
+import { LoadBalancer } from "../components/load_balancer";
+import { Vpc } from "../components/vpc";
 
-const commonTags = {};
+function removeFirstSubdomain(domain: string) {
+  return domain.substring(domain.indexOf('.') + 1);
+}
+
+const config = new pulumi.Config();
+
+const projectName = pulumi.getProject();
+const stackName = pulumi.getStack();
+const prefix = `${projectName}-${stackName}`;
+
+const domainName = config.require("domainname"); // "example1.dev.wci-test.org";
+const zoneName = removeFirstSubdomain(domainName); // "dev.wci-test.org"
+
+const commonTags = {
+  project: projectName,
+  stack: stackName
+};
 
 const appImage = new Image('wci-hello', {
   folder: "../app",
@@ -14,23 +29,28 @@ const appImage = new Image('wci-hello', {
   tags: { ...commonTags }
 })
 
+const myvpc = new Vpc("wci-vpc");
 
-const cluster = new Cluster('wci-dev', {
-  vpc: vpc.vpc,
-  //tags: { "foo": "bar" };
+const alb = new LoadBalancer("wci-alb", {
+  domainName,
+  zoneName,
+  vpc: myvpc.vpc,
+  publicSubnetIds: pulumi.output(myvpc.publicSubnets.map(s => s.id))
 });
 
-const logGroup = new aws.cloudwatch.LogGroup('logs');
+const cluster = new Cluster('wci-dev', {
+  vpc: myvpc.vpc,
+  tags: commonTags
+});
 
-const service = new Service('http', {
+new Service('http', {
   cluster: cluster.cluster,
   taskExecutionRole: cluster.ecsTaskExecutionRole,
-  vpc: vpc.vpc,
+  vpc: myvpc.vpc,
   alb: alb,
-  app: { image: appImage.image, imageRef: appImage.image.ref },
-  publicSubnets: vpc.publicSubnets,
-  logGroup: logGroup,
-  tags: { "magic": "foo" }
+  app: { image: appImage.image },
+  publicSubnets: myvpc.publicSubnets,
+  tags: commonTags
 })
 
-export const loadBalancerUrl = 'https://' + alb.albDomainName;
+export const loadBalancerUrl = 'https://' + alb.domainName;
